Updated Table Names
Updated the table names so that we will be able to have a player and
teams table for each sport. Fixed possible circular import error by
importing the models in the init file in models.
