+++ backend/app/db/models/__init__.py
@@ -1,0 +1,3 @@
+from .league import League
+from .sport import Sport
+from .sports.nba import NBAPlayer, NBATeam
