+++ backend/app/scripts/nba/fetch_players.py
@@ -1,8 +1,7 @@
 import asyncio
 
 from app.db.database import async_engine
-from app.db.models.sports.models import Team
-from app.db.models.nba.models import Player
+from app.db.models.sports.nba import NBATeam, NBAPlayer
 from app.utils.fetch_utils import fetch_data_async
 from nba_api.stats.endpoints import commonplayerinfo
 from nba_api.stats.static import players
@@ -113,7 +112,7 @@ async def insert_all_player_info(player_info_dfs):
                         )
                         continue
 
-                    result = await session.execute(select(Team).filter_by(id=team_id))
+                    result = await session.execute(select(NBATeam).filter_by(id=team_id))
                     team_exists = result.scalars().one_or_none()
 
                     # Check if the team id tied to the player is valid
@@ -123,7 +122,7 @@ async def insert_all_player_info(player_info_dfs):
                         )
                         continue
 
-                    player = Player(**player_info_dict)
+                    player = NBAPlayer(**player_info_dict)
                     await session.merge(player)
 
         await session.commit()
