+++ backend/app/services/nba_service.py
@@ -1,11 +1,11 @@
-from app.db.models.nba.models import Player
+from app.db.models.sports.nba import NBAPlayer
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 
 def get_players(db: Session):
-    return db.execute(select(Player)).scalars().all()
+    return db.execute(select(NBAPlayer)).scalars().all()
 
 
 def get_player_info(db: Session, player_id: int):
-    return db.execute(select(Player).filter(Player.player_id == player_id)).scalar()
+    return db.execute(select(NBAPlayer).filter(NBAPlayer.player_id == player_id)).scalar()
